+++ patlas/MASHix.py
@@ -202,6 +202,17 @@ def master_fasta(fastas, output_tag, mother_directory):
 
     """
 
+    sequences_not_to_remove = [
+
+    ]
+    """
+    A list with the sequences that shouldn't be removed by the keywords "cds"
+    and "origin". This require that we already know which ones should not be
+    removed. So it is a good idea to let the script run until this step first
+    each time the database is updated. Using the search 
+    "--search-sequences-to-remove" option.
+    """
+
     # initiates output for statistics of removed files
     remove_seq_out = open(os.path.join(mother_directory,
                                        "removed_entries_stats.txt"), "w")
@@ -263,12 +274,30 @@ def master_fasta(fastas, output_tag, mother_directory):
                             # after appending new length to dicts reset lengths
 
                         if not truePlasmid:
-                            remove_seq_out.write("\t".join([
-                                previous_header.replace("\n", ""),
-                                str(length),
-                                reason,
-                                filename
-                            ]) + "\n")
+                            # force sequence to be added although truePlasmid
+                            # is false. This because it is in a list that will
+                            # bypass this filter
+                            if sequences_not_to_remove:
+                                sequence_info[accession] = (species, length,
+                                                           plasmid_name)
+                                # dict at the beginning of each new entry
+                                master_fasta.write(previous_header)
+                                master_fasta.write("".join(previous_sequence))
+
+                                previous_sequence = []
+
+                                length_dict[accession] = length
+
+                            # if it is not in this list then the entry needs to
+                            # be removed from fasta file and therefore from
+                            # patlas.
+                            else:
+                                remove_seq_out.write("\t".join([
+                                    previous_header.replace("\n", ""),
+                                    str(length),
+                                    reason,
+                                    filename
+                                ]) + "\n")
 
                         # resets sequence length for every >
                         length = 0
@@ -816,7 +845,6 @@ def mash_distance_matrix(mother_directory, sequence_info, pvalue, mashdist,
             cluster_info = str(clusted_id[0])
         else:
             cluster_info = None
-        #print(doc)
         doc["cluster"] = cluster_info
 
         # finally adds row to database
@@ -894,6 +922,15 @@ def main():
                                required=True, help="specify the path to the "
                                                    "file containing names.dmp "
                                                    "from NCBI")
+    other_options.add_argument("--search-sequences-to-remove",
+                               dest="sequences_to_remove",
+                               action="store_true",
+                               help="this option allows to only run the part "
+                                    "of the script that is required to "
+                                    "generate the filtered fasta. Allowing for "
+                                    "instance to debug sequences that shoudn't "
+                                    "be removed using 'cds' and 'origin' "
+                                    "keywords")
 
     args = parser.parse_args()
 
@@ -925,7 +962,15 @@ def main():
     main_fasta, sequence_info, all_species = master_fasta(fastas, output_tag,
                                              mother_directory)
 
-    return
+
+    # if the parameter sequences_to_remove is provided the script will only
+    # generate the fasta files and a list of the sequences that were removed
+    # from ncbi refseq original fasta.
+    if args.sequences_to_remove:
+        print("\nDebug mode for searching sequences to remove enabled! "
+              "Leaving script...")
+        sys.exit(0)
+
     #########################
     ### genera block here ###
     #########################